r393: This is a large patch to add support for YUV4MPEG streams.  This will allow Cinelerra to open pipes to the mjpegtools toolchain, such as mpeg2enc.  Files can be opened as well, but not via pipes.
This patch is large enough that it may well introduce some bugs,
possibly not only in YUV export, but in other rendering areas, and
possibly autoconf.  If this happens, I'll fix things as fast as I can.

This module introduces a new dependency on libmjpegutils.a, which can
be found at <http://mjpeg.sourceforge.net>.  The autoconf support I
put in is still rather rough, and needs to be revisited.  But it
should be enough for people to start playing with...

(We can always back it out, right?)