r751: Do not use a constant build date; mark this version as 'CV'.
The build date is actually set at configure time, not at build time,
but this is better than the constant time that was inherited from the official
release.

Additionally, this verison of Cinelerra is marked as 'CV' to distinguish it
from the official version. This should make it much easier to discern
bug reports for vanilla from CV version.