r792: Make sure waveforms are shown on project loading ...
When loading a file that has an index already generated and also includes clips, waveform is not shown on first load (attached b.xml). 

The problem seems to come from difference in loading media as part of clips or as part of main edl. In the former case, the assets keep INDEX_NOTTESTED flag which prevents showing the waveform.

The patch simply makes sure that first the main EDL is loaded and then both the vwindow and clips EDLs...

i get "waveform" not shown bugs reported in person a lot lately, so i hope this is the fix, but there might be additional cause for it also.