r985: Fill in the AUTHORS file with useful information.
This file is used by a svn to git converter; hence, it must have this
particular syntax, which maps svn committer logins to names and emails.

On the other hand, since this file lists only the svn logins, it is by
no means a complete list of authors. Many people have contributed, but
are not mentioned here. If you think you deserve attribution in this
list, it will be no problem to add your name, too (with some fake
login name).