r987: Fix incomplete UI updates and re-rendering after clicking in the time line.
Some updates of the plugin windows and fader slider or missing
re-renderings happen at the following instances:

- Clicking on an edit boundary.
- Clicking on a plugin boundary.
- Jumping from label to label with Ctrl-Arrow.
- Jumping to the beginning or end of the time line with Pos1 or End.
- Selecting a label in the main windows.
- Selecting a label in the composer.
- Clicking on an automation keyframe.

The changes in trackcanvas.C basically unify all do_*() routines to set
the flags that indicate what to change to the caller, because all
callers already were prepared to do the updates, but were not told to do
them.

In the other instances the necessary updates are introduced.

This fixes bug 327.