r1012: Add a new video plugin for rerouting channels to a different track.
The new plugin enables to selectively transfer the Alpha channel
or the Components (RGB or YUV) or both from a source track to
a target track, partially overwriting the target's contents.
It works as a shared plugin. The typical usage scenario is
to build up a possibly animated Mask in one track and then
to transfer the Alpha channel to another content track

Authored by Hermann Vosseler.