r1020: Correct colors drawn in vectorscope.
Colors drawn in the vectorscope in the BGR8888 color model had swapped red
and blue. As a result, the display inaccurately reflected the source image,
and adjustments with the Color Balance plug-in were confusing because image
changes did not follow the apparent color axes.

This change insures points are drawn in their correct colors on the correct
color axes.

Patch by Craig Lawson.