r1021: Draw waveform points in color, too.
Vectorscope points are drawn in color and it looks nice. Draw waveform
points in color so they look nice, too. Color waveform helps the user to
visually correlate the frame with the waveform.

Also brighten and decrease contrast of points drawn. Although this distorts
the videoscope colors slightly, if this is not done the very dim colors are
not visible against the black background.

Patch by Craig Lawson.