r1043: PLUGIN:Motion - Ability to add offsets from tracked frame
To stabilize video, the motion plugin uses a "tracking frame" to which
to track to and a region within that frame to track (generally an
object in the background) in the current frame. When the region is
obscured, (often by something in the foreground, or by leaving the
screen), then the motion compensation would fail, and the video jump's
all over the place.

This patch helps work around the problem by allowing one to set a
second region to track and then add offset the (previously calculated and
saved to the respective file in /tmp/???/)

Here's how I use it:

Timeline: k--------k-----------------------
          ^       ^^
          A       BC

A - object1 is visible in the background up until C
B - (the frame before C) has both object1 and object2 visible
C - has only object2 visible

1. Make a keyframe and set to track object1
2. Make a keyframe at C and track frame at B, set to track object2
3. Set keyframe at C to add offsets that were calcuated at B.

Done! And it works REALLY well!!! (I am really happy with this! :P)