Fix a crash in the Delay Audio plugin.
The plugin used a buffer to save away audio samples until as requested by
the configuration. The delay algorithm required a buffer size that can hold
the samples of the delay time plus the samples of the frame that was
processed. However, the value of PluginClient::in_buffer_size was used to
allocate the buffer, but the size parameter of process_realtime() could
be larger, so that samples were written beyond the buffer.

This replaces the whole delay algorithm. The important change is that the
algorithm does not depend on "reconfiguration", but looks at the number of
samples to delay at each invocation and adjusts the buffer accordingly.
The buffer itself is now implemented with a std::vector and std::copy is
used instead of plain memcpy().