Clean up .gitignore files.
Almost every subdirectory had its own .gitignore file with the same content
(Makefile and Makefile.in). This is a heritage from SVN's svn:ignore
property that was converted to git.

This removes the unnecessary repetition and also makes those entries in the
top-level .gitignore more specific that should not apply recursively.