Pass --libdir given to configure on to ffmpeg configure.
Previously, the sub-configure of the internal ffmpeg would always use the
default libdir ($prefix/lib).

The difficulty is that the default $libdir that autoconf generates contains
the Makefile variable ${exec_prefix} (which also happens to be a configure
shell variable). To get its value, $libdir must be expanded twice. A
further complication is that the default of exec_prefix is NONE, which we
have to replace by $prefix.

Based on suggestions by Reimar Imhof.