Do not install ffmpeg's header files.
The internal ffmpeg is not intended to be used for development of third-
party software. We only want its libraries. Do not install the header
files so that an existing ffmpeg installation is not disturbed.

The libraries are already installed with a customized name so that the do
not interfere.