Let more precision digits appear in the ReframeRT Scale value.
On long continuous shots like speeches taken with separate devices
for audio and video, the minimal differences in their internal
clocks will make them drift by a couple of seconds even if
synchronized at the beginning.

This requires a precise calculation in ReframeRT to stretch the
video sources to resync them. 4 digits are not enough.