Fix loading default keyframes from old projects
Old default keyframes are overwrited with te data from
the keyframe positioned at the beginning of the plugin.

Incompatibility with old projects:
If there is no keyframe the beginnig of a plugin the default
keyframe becomes the first keyframe of plugin.
From keyframes that are too near each other (less than track unit),
only the latest will remain.