Editing and dragging of automation nodes and tangents
eworked the handling of drag and edit operations.
 - toggle tangent mode by ctrl-click
 - make the tangent itself draggable by ctrl
 - make the tangent behave like a lever, so you can use it to
   point on every location on canvas.
 - allow to ctrl-click and immediately start dragging the tangent.
 - ALT fixes position of current keyframe. (shift snaps to near by values)

I had some problems with guicast messing up ctrl, alt, and shift callbacks
in a non predictable manner when the mouse moves fast. I couldn't isolate
the source of this, but I tried to make the handling a bit more stable
under this circumstances.

Backport from official branch.