Tangent mode toggles in compositor tools window
This was proposed by Andraz Tori. Add a a shortcut to toggle
the auto smoothing/linear mode for all three camera/projector
automations together. Handle redrawing. The state of the toggles
depends on the tangent mode of the next keyframe to the left
on each of the three curves.

Backport from official branch.