Avoid useless malloc() and memory leak in mplex.c
This patch fixes the issues I found in mplex.c using Coverity and
valgrind.

The mplex.c source contain a useless malloc(1), which seem to be there
to make sure a float pointer can be relased using free() when scaling
the buffer to the required size.  But the C standard specify that
free(NULL) work fine (apparently it crashes on obsolete platforms like
PalmOS and 3BSD, neither which are relevant platforms for Cinelerra), so
it is safe to initialize the audio_temp pointer to NULL instead.  The
audio buffer memory allocated later in the code is never released on
exit.  Neither is the memory allocated to the path string.  Add code to
release it just to clean up the heap, as this is done just before main()
terminates the program, to get rid of valgrind reported leak.  In the
process I made sure to correct the path array size calculation to ignore
argv[0].  Also, move the path allocation after argument checking to
avoid having to release it if the usage() info is printed.