Titler plugin: added automatic conversion to utf8 from other encodings.
Now titler plugin on utf8 converts encodings to utf8 when saves keyframes.
On not utf8 system can handle also utf8 text, obviously with text input
limitation can only render and edit with others char codes. For this reason
I have disabled the auto conversion with not utf8 systems. A new function,
convert_encoding, does an utf8 conversion if the input text is not utf8,
but works in the different way on two systems, on not utf8 systems the
conversion is temporary, in order to work with ucs4 converter. Deleted
old iconv code that not working more.