Remove accidental double assignment
Coverity CID 57787 reports the duplicated assignment to pkg->row2. This
is a remnant of the rewrite in c7737fca (Totaly awsome new, fast and
accurate maskengine). Before that rewrite, the assignment was to two
different packages that were treated per iteration. Since then we have
only one package per loop and the end condition is already taken care
of, and we can remove the if-statement.

Thanks to Nicola Ferralis for an initial patch.