Add destructors to fix resource leaks
The classes SetFormat and SetFormatThread in setformat.C
and RemoveThread in removethread.C do not have destructors,
leaving the system open to resource leaks.

This patch add the corresponding destructors and within each it
adds the proper deletes for deallocating memory for the
varialbles that otherwise would leak.

Found by Coverity.