Fix incorrect decrements of BC_ISlider slider controls.
The value of a slider control can be changed with the mouse wheel and the
arrow buttons: Up and down change in large steps, left and right in small
steps.

The implementation of BC_ISlider, got the decrements the wrong way around,
leading to large increments and small decrements with the up and down
arrow keys and the mouse wheel.