Using intltoolize
This patch fixes all errors and warnings around po/Makefile.in.in and also
we no longer need use --force in libtoolize.

With "new" intltoolize tool, all configure.ac tends to remove
AM_GNU_GETTEXT and add IT_PROG_INTLTOOL

-AM_GNU_GETTEXT([external])
-AM_GNU_GETTEXT_VERSION(0.12.1)
+IT_PROG_INTLTOOL

Finally in autogen.sh, we have to run intltoolize to generate
po/Makefile.in.in, but po/Makefile.in.in already exists in source code.
So, IMO, the best solution is remove po/Makefile.in.in from source code,
because po/Makefile.in.in is an autogenerated file.

Signed-off-by: Sérgio M. Basto <sergio@serjux.com>