Upgrade internal ffmpeg to version 2.8
Configure clones latest realeased version from
source.ffmpeg.org to cinelerra directory.
Cloned version is not upgraded automatically.
The new version is not part of our source tree and
is not stored to the git of cinelerra.
The suffix of ffmpeg libraries is changed to cve,
headers are installed under cinelerra-cve.