Fix implementation of is_root_dir().
The function should check whether argument passed in is the root directory,
but it actually does not look at the parameter at all, but only checks the
member current_dir. Fix this to actually check the parameter. This does not
change the behavior because the function is called only twice and both
callers pass current_dir as argument anyway.