Plugin:Motion - add a gain to the compensation..
The idea was that sometimes there's a blurring that occurs when the
camera shakes.  When motion correction completely fixes the
compensation, the blur looks a bit silly.  The idea was that by
having a gain on the compensation, the video would still move
around, but possibly with %1 of the shake that it had before.. that
way the eye's might correlate the blured direction and ignore it..

Date of original commit: Fri, 21 Dec 2007
Found in arcives by: Igor Vladimirsky <sitelve@gmail.com>