New function Track::get_effects_length
get_effects_length is split out from get_length
The length of a track does not depend on the last plugin
or last auto. It depends on synthesis plugin.
If is_synthesis is zero function returns the end position
of last plugin or position of last auto, whatever is greater