Return the value of need_reconfigure from load_configuration
need_reconfigure is set in send_configure_change when
user twaked the gui of a plugin and means that the plugin
must take some additional action. Return the value from
load_configuration in the case when the change of configuration
is not detected. Clear need_reconfigure after the frame is processed.